requirements.txt
altgraph==0.12
appscript==1.0.1
autotagger==0.1.6
BlinkStick==1.1.8
colorthief==0.2
discogs-client==2.2.1
Django==1.10.5
eyeD3==0.7.10
macholib==1.7
modulegraph==0.12.1
mutagen==1.36
oauthlib==2.0.1
olefile==0.44
Pillow==4.0.0
py-applescript==1.0.0
py2app==0.9
pyobjc==3.0.4
pyusb==1.0.0b1
requests==2.12.4
six==1.10.0
untangle==1.1.0
xmltodict==0.10.1
